% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{albedo}
\alias{albedo}
\title{Calculate canopy albedo}
\usage{
albedo(
  pai,
  lref,
  ltra,
  x,
  gref,
  slope,
  aspect,
  solp,
  scattermethod = "Sellers",
  leafh = FALSE
)
}
\arguments{
\item{pai}{total one side area of canopy elements per unit ground area}

\item{lref}{leaf reflectance (0-1)}

\item{ltra}{leaf transmittance. Note that `ltra + lref` must be < 1.}

\item{x}{ratio of vertical to horizontal projections of leaf foliage}

\item{gref}{ground reflectance}

\item{slope}{slope of ground surface (decimal degrees from horizontal)}

\item{aspect}{aspect of ground surface (decimal degrees from north)}

\item{solp}{a list of solar zenith and azimuths as returned by [sunposition()]}

\item{scattermethod}{one of "Sellers" - leaf directionality factor approximated by
calculating mean leaf angle or "Pinty" - leaf directionality factor calculated
by integrating the inclination density distribution function.}

\item{leafh}{optional logical indicating whether `pai` is the total one-side plant
area per unit horizontal area (TRUE) or per sloped ground surface area (FALSE)}
}
\value{
a list of the following
\describe{
 \item{albd}{combined ground and canopy albedo for diffuse radiation}
 \item{albc}{canopy only albedo for direct radiation}
 \item{albg}{ground only albedo - same as gref}
 \item{albb}{combined ground and canopy albedo for direct radiation}
 \item{wgtg}{relative weighting of ground contribution to total albedo for direct radiation}
 \item{wgti}{relative weighting of ground contribution to total albedo for diffuse radiation}
 \item{K}{canopy extinction coefficient for vegetation above a horizontal surface}
}
}
\description{
Calculates canopy albedos for direct and diffuse radiation using
a two-stream model.
}
