% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal.R
\name{bird_resistance}
\alias{bird_resistance}
\title{Bird heat transfer resistance}
\usage{
bird_resistance(
  height,
  width,
  len,
  rho = 730,
  tbody,
  tair,
  adir,
  atilt,
  position,
  windspeed,
  wdir
)
}
\arguments{
\item{height}{Body height (cm)}

\item{width}{Body width (cm)}

\item{len}{Body length (cm)}

\item{rho}{Body density (kg m^-3). Default is 730.}

\item{tbody}{Body temperature (degrees C)}

\item{tair}{Air temperature (degrees C)}

\item{adir}{body azimuth direction (degrees)}

\item{atilt}{body tilt from vertical (degrees)}

\item{position}{assumption about body orientation. See [chardim()]}

\item{windspeed}{Wind speed at animal height (m s^-1)}

\item{wdir}{wind direction (degrees)}
}
\value{
A list with:
  \describe{
    \item{rHa_bound}{boundary layer resistance (s m^-1)}
    \item{rHa_warm}{Total resistance (s m^-1) under reduced insulation (warm conditions)}
    \item{rHa_cold}{Total resistance (s m^-1) under increased insulation (cold conditions)}
  }
}
\description{
Calculates total resistance to heat loss for a bird, combining boundary
layer resistance with plumage conductance under minimum and maximum
insulation states.
}
\details{
Boundary layer resistance is calculated from body geometry and wind speed.
Plumage conductance is estimated from body mass using an allometric
relationship for minimum conductance, with a higher conductance representing
reduced insulation. Conductance is converted to resistance and adjusted for
wind speed before being combined with boundary layer resistance.
}
