% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentheat.R
\name{bulkstomatalresist_calc}
\alias{bulkstomatalresist_calc}
\title{Calculate bulk surface stomatal resistance}
\usage{
bulkstomatalresist_calc(
  solp,
  Ca,
  Rsw,
  Rdif,
  tair,
  tcanopy,
  rh,
  pk,
  psi_r,
  plant_inputs,
  C3 = TRUE
)
}
\arguments{
\item{solp}{a list of solar positions as returned by [sunposition()]}

\item{Ca}{atmospheric carbon dioxide concentration (ppm)}

\item{Rsw}{flux density of shortwave radiation on the horizontal (W/m^2)}

\item{Rdif}{flux density of diffuse radiation on the horizontal (W/m^2)}

\item{tair}{temperature of air (deg C)}

\item{rh}{relative humidity of air (percentage)}

\item{pk}{atmospheric pressure (kPa)}

\item{psi_r}{# mean water potential in root zone (MPa)}

\item{plant_inputs}{a vector of vegetation parameters as returned by [createplant_inputs()]}

\item{C3}{optional logical indicating whether vegetation has C3 or C4 photosynthetic pathway}

\item{PAI}{total one-sided plant area per unit ground area (m^2 / m^2)}

\item{tleaf}{temperature of leaf (deg C)}
}
\value{
bul surface stomatal resistance (s/m)
}
\description{
Calculate bulk surface stomatal resistance
}
