% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{calcRabs}
\alias{calcRabs}
\title{Calculates radiation absorbed by canopy}
\usage{
calcRabs(climdata, vegp, groundp, lat, long)
}
\arguments{
\item{climdata}{a data.frame of weather variables - see e.g. example dataset micropoint::climdata}

\item{vegp}{a list of vegetation parameters - see e.g. example dataset micropoint::vegparams}

\item{groundp}{a list of ground parameters - see e.g. micropoint::groundparams}

\item{lat}{latitude (decimal degrees), negative south of the equator}

\item{long}{longitude (decimal degrees), negative west of the Greenwich meridian}
}
\value{
radiation absorption by canopy (W/m^2)
}
\description{
Calculates combined longwave and shortwave radiation absorption.
}
