% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{Campbellparams}
\alias{Campbellparams}
\title{Create list of Campbell soil hydraulic parameters}
\usage{
Campbellparams(soiltype)
}
\arguments{
\item{soiltype}{Character string giving the soil type. See details}
}
\value{
A list of soil parameters:
\describe{
  \item{psie}{Air-entry water potential (m)}
  \item{n}{Pore size distribution parameter (unitless)}
  \item{b}{Campbell soil water retention parameter (unitless)}
  \item{Smax}{Volumetric soil water content at saturation (m^3 m^-3)}
  \item{Smin}{Residual volumetric soil water content (m^3 m^-3)}
  \item{Ksat}{Saturated hydraulic conductivity (kg s / m^3)}
}
}
\description{
Create list of Campbell soil hydraulic parameters
}
\details{
:
Available soil types (argument \code{soiltype}) are:
\itemize{
  \item Sand
  \item Loamy sand
  \item Sandy loam
  \item Loam
  \item Silt loam
  \item Sandy clay loam
  \item Clay loam
  \item Silty clay loam
  \item Sandy clay
  \item Silty clay
  \item Clay
  \item Silt
}
}
