% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{cank}
\alias{cank}
\title{Calculates canopy extinction coefficient}
\usage{
cank(solp, x, si)
}
\arguments{
\item{solp}{a list of solar zenith and azimuths as returned by [sunposition()]}

\item{x}{ratio of vertical to horizontal projections of leaf foliage}

\item{si}{the fraction of direct ebam radiation incident on an inclined surface
as returned by [solarindex()]}
}
\value{
a list of the following
\describe{
 \item{k}{canopy extinction coefficient for vegetation above a flat surface}
 \item{k0}{canopy extinction coefficient when solar zenith = 0}
 \item{kd}{canopy extinction coefficient for vegetation above an inclined surface}
}
}
\description{
Calculates the canopy extinction coefficient for inclined and flat canopies
}
