% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensibleheat.R
\name{canopyresistance}
\alias{canopyresistance}
\title{Calculate resistance to sensible heat loss of entire canopy}
\usage{
canopyresistance(z, d, zm, uf, psih)
}
\arguments{
\item{z}{height to which resistance is needed (m)}

\item{d}{zero-plane displacement height as returned by [zeroplanedis()] (m)}

\item{zm}{roughness length for momentum as returned by [roughlength()] (m)}

\item{uf}{wind friction velocity as returned by [windfric()] (m/s)}

\item{psih}{diabatic correction for heat as returned by [dpsih()]}
}
\value{
bulk surface aerodynamic resistance (m/s)
}
\description{
Calculate resistance to sensible heat loss of entire canopy
}
