% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal.R
\name{chardim}
\alias{chardim}
\title{Calculates a characteristic dimension based on projected area}
\usage{
chardim(wdir, height, width, len, adir = 0, atilt = 0, position = "fixed")
}
\arguments{
\item{wdir}{wind direction (degrees)}

\item{height}{body height along the vertical axis (cm)}

\item{width}{body width along the lateral axis (cm)}

\item{len}{body length along the longitudinal axis (cm)}

\item{adir}{body azimuth direction (degrees). Only used when
\code{position = "fixed"}. Default is 0.}

\item{atilt}{body tilt from vertical (degrees). Only used when
\code{position = "fixed"}. Default is 0.}

\item{position}{assumption about body orientation. Must be one of:
\item{\code{"fixed"}}{body orientation is fixed according to
  \code{adir} and \code{atilt}}
  \item{\code{"max"}}{body orientation is assumed to maximise projected
  silhouette area}
  \item{\code{"min"}}{body orientation is assumed to minimise projected
  silhouette area}
  \item{\code{"randomdir"}}{body azimuth direction is assumed random, but
  tilt is fixed by \code{atilt}}
  \item{\code{"random"}}{both body azimuth direction and tilt are assumed
  random}}
}
\value{
a numeric value giving the characteristic dimension (m), defined as
volume divided by projected silhouette area
}
\description{
Computes a characteristic dimension of an ellipsoidal body as
the ratio of volume to projected silhouette area in the direction of wind
flow. The projected area is obtained from [silhouette()], with body
orientation specified directly or inferred from simple assumptions.
}
\details{
The characteristic dimension is calculated as:
\deqn{d = V / A_p}
where \eqn{V} is ellipsoid volume and \eqn{A_p} is projected area normal to
the flow direction. The projected area is always evaluated for a horizontal
beam (i.e. \eqn{90^\circ} zenith angle) aligned with the wind direction.
Solar angles (\code{zen}, \code{azi}) are only used to determine body
orientation when \code{position} is set to \code{"max"} or \code{"min"}.
}
