% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{clearskyrad}
\alias{clearskyrad}
\title{Clear-sky solar radiation}
\usage{
clearskyrad(climdata, lat, long)
}
\arguments{
\item{climdata}{Data frame containing at least \code{obs_time},
\code{temp}, \code{relhum} and \code{pres}.}

\item{lat}{Latitude (decimal degrees)}

\item{long}{Longitude (decimal degrees)}
}
\value{
Numeric vector of clear-sky radiation (W m^-2)
}
\description{
Calculates clear-sky direct solar radiation from time, location, air
temperature, relative humidity and pressure.
}
