% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensibleheat.R
\name{cpair}
\alias{cpair}
\title{Calculate specific heat of air at constant pressure}
\usage{
cpair(tc)
}
\arguments{
\item{tc}{air temperature (degrees C)}
}
\value{
specific heat of air at constant pressure (J/mol/K)
}
\description{
Calculate specific heat of air at constant pressure
}
