% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentheat.R
\name{createplant_inputs}
\alias{createplant_inputs}
\title{Creates a list of vegetation parameter inputs for a given plant function type}
\usage{
createplant_inputs(PFT)
}
\arguments{
\item{PFT}{a plant function type as listed in `PFTtable`}
}
\value{
a list of vegetation parameters.
}
\description{
Creates a list of vegetation parameter inputs for a given plant function type
}
