% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentheat.R
\name{dewpoint}
\alias{dewpoint}
\title{Dew-point temperature}
\usage{
dewpoint(tc, rh)
}
\arguments{
\item{tc}{Air temperature (°C).}

\item{rh}{Relative humidity (%).}
}
\value{
Dew-point temperature (°C).
}
\description{
Calculate dew-point temperature (°C) from air temperature (°C) and
relative humidity (%). Uses Magnus equations with coefficients for
water and ice, applying the ice formulation when dew point is below 0 °C.
}
\seealso{
[satvap()]
}
