% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{difprop}
\alias{difprop}
\title{Diffuse proportion of shortwave radiation}
\usage{
difprop(climdata, lat, long)
}
\arguments{
\item{climdata}{Data frame containing \code{obs_time} and \code{swdown}.}

\item{lat}{Latitude (decimal degrees)}

\item{long}{Longitude (decimal degrees)}
}
\value{
Numeric vector giving the diffuse proportion of shortwave radiation
}
\description{
Calculates the diffuse fraction of incoming shortwave radiation from total
downward shortwave flux and solar geometry.
}
