% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow.R
\name{.appendSnowSoil}
\alias{.appendSnowSoil}
\title{Internal functions used by snow-soil heat solver}
\usage{
.appendSnowSoil(snowmod, soilheatmod, soilp, atmPressure)
}
\description{
Internal functions used by snow-soil heat solver
}
