% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{.dvapor_dPsi}
\alias{.dvapor_dPsi}
\title{Internal function for used my soil water model}
\usage{
.dvapor_dPsi(psiw, psie, theta, b, Smax, Tk)
}
\description{
Internal function for used my soil water model
}
