% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentheat.R
\name{.photosynth_model}
\alias{.photosynth_model}
\title{Internal plant photosynthesis model}
\usage{
.photosynth_model(
  Ca,
  Rswabs,
  tair,
  tleaf,
  rh,
  pk,
  psi_r,
  plant_inputs,
  C3 = TRUE
)
}
\description{
Internal plant photosynthesis model
}
