% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{.root_distribute}
\alias{.root_distribute}
\title{Internal function used to distribute roots zone}
\usage{
.root_distribute(dz, totalDepth, skew)
}
\description{
Internal function used to distribute roots zone
}
