% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentheat.R
\name{.rpmin_calc}
\alias{.rpmin_calc}
\title{internal function for calculating minimum resistance}
\usage{
.rpmin_calc(h, hv = 0.002, Kxmx)
}
\description{
internal function for calculating minimum resistance
}
