% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{.transpiration_distribute}
\alias{.transpiration_distribute}
\title{Internal function used to distribute transpiration in root zone}
\usage{
.transpiration_distribute(soilp, rootfrac, totalTransp_mm, dT, psiw, p = 0.5)
}
\description{
Internal function used to distribute transpiration in root zone
}
