% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensibleheat.R
\name{dphih}
\alias{dphih}
\title{Calculate diabatic influencing coefficient for heat}
\usage{
dphih(ze)
}
\arguments{
\item{ze}{atmospheric stability parameter}
}
\value{
diabatic influencing coefficient for heat
}
\description{
Calculate diabatic influencing coefficient for heat
}
