% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensibleheat.R
\name{dpsih}
\alias{dpsih}
\title{Calculate integrated diabatic correction coefficient for heat}
\usage{
dpsih(ze)
}
\arguments{
\item{ze}{atmospheric stability parameter}
}
\value{
diabatic correction coefficient for heat
}
\description{
Calculate integrated diabatic correction coefficient for heat
}
