% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensibleheat.R
\name{dpsim}
\alias{dpsim}
\title{Calculate integrated diabatic correction coefficient for momentum}
\usage{
dpsim(ze)
}
\arguments{
\item{ze}{atmospheric stability parameter}
}
\value{
diabatic correction coefficient for momentum
}
\description{
Calculate integrated diabatic correction coefficient for momentum
}
