% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal.R
\name{Ectothermtemp}
\alias{Ectothermtemp}
\title{Calculates ectotherm body temperature over a single timestep}
\usage{
Ectothermtemp(
  ecto_params,
  tme,
  microcdata,
  lat,
  long,
  tolerance = 0.001,
  maxiter = 100
)
}
\arguments{
\item{ecto_params}{A list of ectotherm parameters as returned by
[micropoint::create_ectop()]}

\item{tme}{POSIXlt object indicating date and time in UTC}

\item{microcdata}{A list of microclimate variables with the following entries:
\describe{
  \item{Tair}{Air temperature (degrees C).}
  \item{Tsurf}{Surface temperature (degrees C).}
  \item{Rdirdown}{Direct shortwave radiation flux perpendicular to the solar beam (W m^-2).}
  \item{Rdifdown}{Diffuse downward shortwave radiation flux (W m^-2).}
  \item{Rswup}{Upward shortwave radiation flux (W m^-2).}
  \item{Rlwdown}{Downward longwave radiation flux (W m^-2).}
  \item{Rlwup}{Upward longwave radiation flux (W m^-2).}
  \item{winddir}{Wind direction (degrees from north).}
  \item{windspeed}{Wind speed at animal height (m s^-1).}
  \item{relhum}{Air relative humidity (Percentage).}
  \item{pk}{Atmospheric pressure (kPa).}
  \item{surfrelhum}{Relative humidity at the surface (Percentage).}
}}

\item{lat}{Latitude (decimal degrees).}

\item{long}{Longitude (decimal degrees).}

\item{tolerance}{Convergence tolerance used when iterating (degrees C).}

\item{maxiter}{Maximum number of iterations.}
}
\value{
Body temperature (deg C)
}
\description{
Calculates ectotherm body temperature over a single timestep
}
