% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal.R
\name{Endothermstress}
\alias{Endothermstress}
\title{Endotherm energy balance}
\usage{
Endothermstress(endo_params, tme, microcdata, lat, long)
}
\arguments{
\item{endo_params}{A list of endotherm parameters with the following entries:
\describe{
  \item{height}{Body height (cm).}
  \item{width}{Body width (cm).}
  \item{len}{Body length (cm).}
  \item{refl}{Shortwave reflectance of the body surface (unitless).}
  \item{em}{Longwave emissivity of the body surface (unitless).}
  \item{confrac}{Fraction of body surface in conductive contact with the substrate (unitless).}
  \item{adir}{Body orientation relative to north (degrees).}
  \item{atilt}{Body tilt angle relative to the horizontal (degrees).}
  \item{position}{Descriptor of body position (character).}
  \item{Tbody}{Body temperature (degrees C).}
  \item{type}{Endotherm type; one of \code{"mammal"} or \code{"bird"}.}
  \item{basmet}{Basal metabolic rate (W).}
  \item{coatthick}{Coat thickness (cm; mammals only).}
  \item{rho}{Body density (kg m^-3; birds only).}
}}

\item{tme}{POSIXlt object indicating date and time in UTC.}

\item{microcdata}{A list of microclimate variables with the following entries:
\describe{
  \item{Tair}{Air temperature (degrees C).}
  \item{Tsurf}{Surface temperature (degrees C).}
  \item{Rdirdown}{Direct shortwave radiation flux perpendicular to the solar beam (W m^-2).}
  \item{Rdifdown}{Diffuse downward shortwave radiation flux (W m^-2).}
  \item{Rswup}{Upward shortwave radiation flux (W m^-2).}
  \item{Rlwdown}{Downward longwave radiation flux (W m^-2).}
  \item{Rlwup}{Upward longwave radiation flux (W m^-2).}
  \item{winddir}{Wind direction (degrees from north).}
  \item{windspeed}{Wind speed at animal height (m s^-1).}
  \item{pk}{Atmospheric pressure (kPa).}
}}

\item{lat}{Latitude (decimal degrees).}

\item{long}{Longitude (decimal degrees).}
}
\value{
Net energy balance (W), with positive values indicating heat gain
  and negative values indicating heat loss.
}
\description{
Calculates net energy balance for an endotherm at a single time step from
body traits and local microclimate.
}
\details{
Absorbed shortwave and longwave radiation are calculated from body geometry,
orientation and surface properties. Emitted longwave radiation is determined
from body temperature. Sensible heat exchange is calculated using resistances
appropriate for mammals or birds, representing different insulation states.
Basal metabolic heat production is added and the resulting energy balance is
returned for the whole animal.
}
