% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance.R
\name{energybalance}
\alias{energybalance}
\title{Calculate energy balance}
\usage{
energybalance(Ta, Ts, pk, rh, Rabs, rHa, rS, G = 0, em = 0.97)
}
\arguments{
\item{Ta}{Air temperature (degrees C)}

\item{Ts}{temperature of surface for which energy balance is computed (degrees C)}

\item{pk}{atmospheric pressure (kPa)}

\item{rh}{relative humidity (percentage)}

\item{Rabs}{flux density of absorbed radiation (W/m^2)}

\item{rHa}{resistance to heat loss (s/m)}

\item{rS}{stomatal resistance (s/m)}

\item{G}{rate of heat storage (W/m^2)}

\item{em}{emissivity of surface (unitless)}
}
\value{
energy balance (W/m^2).
}
\description{
Calculate energy balance
}
