% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{evaporation_flux}
\alias{evaporation_flux}
\title{Calculates soil evaporation flux}
\usage{
evaporation_flux(theta, Tsurface, Tair, relhum, rHa, psie, b, Smax, dT = 3600)
}
\arguments{
\item{theta}{volumetric soil surface water content (m3 m-3)}

\item{Tsurface}{soil surface temperature (degrees C)}

\item{Tair}{air temperature (degrees C)}

\item{relhum}{air relative humidity (Percentage)}

\item{rHa}{aerodynamic resistance to heat as returned by [groundresistance()] (s m-1)}

\item{psie}{air-entry water potential (J kg-1)}

\item{b}{Campbell soil water retention parameter}

\item{Smax}{volumetric soil water content at saturation (m3 m-3)}

\item{dT}{time step (s)}
}
\value{
soil evaporation (mm over timestep)
}
\description{
Calculates soil evaporation flux
}
