% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensibleheat.R
\name{groundresistance}
\alias{groundresistance}
\title{Calculate resistance from ground surface to some height zref above canopy}
\usage{
groundresistance(hgt, pai, uf, LL, psih, zref)
}
\arguments{
\item{hgt}{height of canopy (m)}

\item{pai}{plant area index}

\item{uf}{wind friction velocity as returned by [windfric()] (m/S)}

\item{LL}{Obukhov length as returned by [Obukhov()]}

\item{psih}{diabatic influencing factor for heat as returned by [dpsih()]}

\item{zref}{height to which resistance is calculated (m)}
}
\value{
resistance to heat loss (s/m)
}
\description{
Calculate resistance from ground surface to some height zref above canopy
}
