% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{heatCapacity}
\alias{heatCapacity}
\title{Calculates soil volumetric heat capacity}
\usage{
heatCapacity(Vq, Vm, Vo, Vw, Tc, pk)
}
\arguments{
\item{Vq}{volumetric quartz fraction of soil (m^3 / m^3)}

\item{Vm}{volumetric mineral fraction of soil (m^3 / m^3)}

\item{Vo}{volumetric organic fraction of soil (m^3 / m^3)}

\item{Vw}{volumetric water content of soil (m^3 / m^3)}

\item{Tc}{soil temperature (degrees C)}

\item{pk}{air pressure (kPa)}
}
\value{
soil volumetric heat capacity (J m-3 K-1)
}
\description{
Calculates soil volumetric heat capacity (J m-3 K-1)
}
