% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{hydraulicConductivityFromTheta}
\alias{hydraulicConductivityFromTheta}
\title{Calculates hydraulic conductivity from soil water content using Campbell method}
\usage{
hydraulicConductivityFromTheta(theta, b, Smax, Ksat)
}
\arguments{
\item{theta}{volumetric soil water content (m3 m-3)}

\item{b}{Campbell soil water retention parameter}

\item{Smax}{volumetric soil water content at saturation (m3 m-3)}

\item{Ksat}{saturated hydraulic conductivity (kg s / m^3)}

\item{psie}{air-entry water potential (J kg-1)}

\item{n}{Campbell pore size distribution parameter}
}
\value{
hydraulic conductivity (kg s / m^3)
}
\description{
Calculates hydraulic conductivity from soil water content using Campbell method
}
