% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow.R
\name{InitSnowModel}
\alias{InitSnowModel}
\title{Initialise a multilayer snow model}
\usage{
InitSnowModel(snowdepth, snowage, snowtemp, snowenv = "Maritime")
}
\arguments{
\item{snowdepth}{Total snow depth (m).}

\item{snowage}{Snow age (hours).}

\item{snowtemp}{Initial snow temperature (\eqn{^\circ}C). This may be a
single value, in which case all layers are assigned the same temperature.
If a vector is supplied, it must have length equal to the number of snow
layers generated.}

\item{snowenv}{Character string specifying the snow environment passed to
[SnowDensParams()]. Options depend on that function, i.e. `Maritime`,
`Alpine`, `Prairie`, `Tundra` or `Taiga`.}
}
\value{
A list containing the snow model state:
\describe{
  \item{SWE}{Snow water equivalent in each layer (mm)}
  \item{depth}{Thickness of each snow layer (m)}
  \item{rho}{Bulk density of each layer (kg m^-3)}
  \item{temp}{Temperature of each layer (deg C)}
  \item{oldtemp}{Temperature of each layer in previous time step (same as temp here)}
  \item{age}{Age of each layer (hours)}
  \item{melt}{Total melt produced over the timestep (mm)}
  \item{snowfall}{Snowfall added over the timestep (mm)}
  \item{rainfall}{Rainfall added over the timestep (mm)}
  \item{throughfall}{Liquid water leaving the base of the snowpack (mm)}
  \item{surfaceWater}{Temporary liquid water stored in the snowpack (mm)}
}
}
\description{
Initialises a multilayer snowpack from total snow depth, snow
age, and snow temperature.
}
