% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{InitSoilheatmod}
\alias{InitSoilheatmod}
\title{Initalizes input for the soil heat model}
\usage{
InitSoilheatmod(soilp, Te, wc)
}
\arguments{
\item{soilp}{list of soil parameters as returned by [createsoilplist()]}

\item{Te}{vector of soil layer temperatures (degrees C)}

\item{wc}{vector of volumetric soil water contents (m3 m-3)}
}
\value{
a list comprising
\describe{
  \item{n}{number of soil layers}
  \item{z}{depth of soil nodes (m)}
  \item{dz}{soil layer thicknesses (m)}
  \item{zCenter}{depth of layer centres (m)}
  \item{vol}{soil layer volumes per unit ground area (m)}
  \item{wc}{volumetric soil water content of each layer (m3 m-3)}
  \item{Te}{soil layer temperature (degrees C)}
  \item{oldTe}{soil layer temperature at previous time step (degrees C)}
  \item{Gflux}{soil heat flux}
  \item{iters}{number of solver iterations}
}
}
\description{
Creates the soil state object used by the soil heat model
}
