% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{InitSoilwatermod}
\alias{InitSoilwatermod}
\title{Initializes input for the soil water model}
\usage{
InitSoilwatermod(soilp, soilheatmod, rootskew)
}
\arguments{
\item{soilp}{list of soil parameters as returned by [createsoilplist()]}

\item{soilheatmod}{soil heat model state as returned by [InitSoilheatmod()]}

\item{rootskew}{parameter controlling vertical distribution of roots}
}
\value{
a list comprising
\describe{
  \item{n}{number of soil layers}
  \item{z}{depth of soil nodes (m)}
  \item{dz}{soil layer thicknesses (m)}
  \item{zCenter}{depth of layer centres (m)}
  \item{Tc}{soil temperature profile (degrees C)}
  \item{theta}{volumetric soil water content (m3 m-3)}
  \item{vol}{soil layer volumes per unit ground area (m)}
  \item{psiw}{soil water potential (J kg-1)}
  \item{k}{hydraulic conductivity (m s-1)}
  \item{vapor}{soil vapour content}
  \item{oldvapor}{soil vapour content at previous time step}
  \item{oldtheta}{soil water content at previous time step (m3 m-3)}
  \item{rootfrac}{fraction of roots in each soil layer}
  \item{success}{logical indicating convergence success - set to TRUE on initalization}
  \item{iterations}{number of solver iterations - set to zero initially}
}
}
\description{
Creates the soil state object used by the soil water model from
an existing soil heat model state
}
