% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentheat.R
\name{latvap}
\alias{latvap}
\title{Calculate latent heat of vapourisation (J / mol)}
\usage{
latvap(Ts)
}
\arguments{
\item{Ts}{surface temperature (degrees C)}
}
\value{
latent heat of vapourisation (J / mol)
}
\description{
Calculate latent heat of vapourisation (J / mol)
}
