% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensibleheat.R
\name{leafresistance}
\alias{leafresistance}
\title{Calculate resistance to sensible heat loss of individual canopy elements}
\usage{
leafresistance(tair, dT, uz, len, wid, x, rHmax = 300)
}
\arguments{
\item{tair}{air temperature (deg C)}

\item{dT}{difference between leaf and air temperature (deg C)}

\item{uz}{wind speed at height of leaf (m/S)}

\item{len}{leaf length (m)}

\item{wid}{leaf width (m)}

\item{x}{ratio of vertical to horizontal projections of leaf foliage}

\item{rHmax}{maximum resistance (used to ensure stable convergence (s/m))}
}
\value{
leaf boundary layer resistance
}
\description{
Calculate resistance to sensible heat loss of individual canopy elements
}
