% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal.R
\name{metabolic_rate}
\alias{metabolic_rate}
\title{Calculates metabolic heat production as a function of body mass and temperature.}
\usage{
metabolic_rate(volume, rho, Q10, a0, b, Tref, Tbody)
}
\arguments{
\item{volume}{Animal volume (m^3)}

\item{rho}{Animal density (kg m^-3)}

\item{Q10}{Temperature sensitivity coefficient}

\item{a0}{Normalisation constant}

\item{b}{Mass-scaling exponent}

\item{Tref}{Reference temperature (°C)}

\item{Tbody}{Body temperature (°C)}
}
\value{
Numeric vector of metabolic rate (W)
}
\description{
Calculates metabolic heat production as a function of body mass and temperature.
}
