% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensibleheat.R
\name{Obukhov}
\alias{Obukhov}
\title{Calculate Obukhov length}
\usage{
Obukhov(tc, pk, uf, H)
}
\arguments{
\item{tc}{air temperature (degrees C)}

\item{pk}{atmospheric pressure (kPa)}

\item{uf}{wind friction velocity (m/s)}

\item{H}{sensible heat flux (W/m^2)}
}
\value{
Obukhov length (m)
}
\description{
Calculate Obukhov length
}
