% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance.R
\name{PenmanMonteith}
\alias{PenmanMonteith}
\title{Calculate temperature of surface using Penman-Monteith equation}
\usage{
PenmanMonteith(Ta, pk, rh, Rabs, rHa, rS, G, em = 0.97, iters = 4)
}
\arguments{
\item{Ta}{Air temperature (degrees C)}

\item{pk}{atmospheric pressure (kPa)}

\item{rh}{relative humidity (percentage)}

\item{Rabs}{flux density of absorbed radiation (W/m^2)}

\item{rHa}{resistance to heat loss (s/m)}

\item{rS}{stomatal resistance (s/m)}

\item{G}{rate of heat storage (W/m^2)}

\item{em}{emissivity of surface (unitless)}

\item{iters}{integer specifying how many times to iterate the equation (4 achieves near exact solution)}
}
\value{
temperature of surface (deg C)
}
\description{
Calculate temperature of surface using Penman-Monteith equation
}
