% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PFTtable}
\alias{PFTtable}
\title{Plant functional type parameter values}
\format{
## `PFTtable`
A data frame with 26 rows and 13 columns:
\describe{
  \item{varname}{Short-hand name of vegetation parameter}
  \item{Description}{Long-hand description of vegetation parameter}
  \item{units}{units used}
  \item{BET.Tr}{Parameter values for Broadleaf evergreen tropical trees}
  \item{BET.Te}{Parameter values for Broadleaf evergreen temperate trees}
  \item{BDT}{Parameter values for Broadleaf deciduous trees}
  \item{NET}{Parameter values for Needle-leaf evergreen trees}
  \item{NDT}{Parameter values for Needle-leaf deciduous trees}
  \item{C3}{Parameter values for C3 grasses}
  \item{C4}{Parameter values for C4 grasses}
  \item{Esh}{Parameter values for Evergreen shrubs}
  \item{Dsh}{Parameter values for Deciduous shrubs}
  \item{multiplier}{Multiplier used to convert form values in table to values used in model}
  ...
}
}
\source{
<https://gmd.copernicus.org/articles/11/2857/2018/gmd-11-2857-2018.pdf>
}
\usage{
PFTtable
}
\description{
Vegetation parameter values for main plant function types
}
\keyword{datasets}
