% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{PsiFromtheta}
\alias{PsiFromtheta}
\title{Derives soil water potential from volumetric water content using the Campbell method}
\usage{
PsiFromtheta(theta, psie, b, Smax, MPa = TRUE)
}
\arguments{
\item{theta}{volumetric soil water fraction (m^3 / m^3)}

\item{b}{Campbell soil water retention parameter}

\item{Smax}{volumetric soil water fraction at saturation (m^3 / m^3)}

\item{MPa}{optional logical indicating whether to return outputs in Mpa (TRUE) or kPa (FALSE)}

\item{psi_e}{air entry water potential (J/m^3)}
}
\value{
soil water potential (MPa or kPa ~ J/Kg)
}
\description{
Derives soil water potential from volumetric water content using the Campbell method
}
