% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{PsiFromthetaVG}
\alias{PsiFromthetaVG}
\title{Derives soil water potential from volumetric water content using the Van Genuchten method}
\usage{
PsiFromthetaVG(theta, Smin, Smax, psie, alpha, n)
}
\arguments{
\item{theta}{volumetric soil water fraction (m^3 / m^3)}

\item{Smin}{Residual volumetric soil water fraction (m^3 / m^3)}

\item{Smax}{volumetric soil water fraction at saturation (m^3 / m^3)}

\item{psie}{air entry water potential (J/m^3)}

\item{alpha}{Van Genuchten shape parameter}

\item{n}{Van Genuchten pore size distribution parameter}
}
\value{
soil water potential (kPa ~ J/Kg)
}
\description{
Derives soil water potential from volumetric water content using the Van Genuchten method
}
