% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentheat.R
\name{relhum_below}
\alias{relhum_below}
\title{Calculate relative humidity below canopy}
\usage{
relhum_below(hgt, paii, TL, uf, pk, Lz, rh, tair, tleaf, tground, soilrh)
}
\arguments{
\item{hgt}{height of canopy(m)}

\item{paii}{vector of plant area index values per unit ground area ordered
sequentially from canopy bottom to top.}

\item{TL}{Langrangian timescale as returned by [solve_wholecanopy()]}

\item{uf}{wind friction velocity as returned by [windfric()] (m/s)}

\item{pk}{atmospheric pressure (kPa)}

\item{Lz}{vector of latent heat fluxes from each canopy layer (W/m^2)}

\item{rh}{vector of initial estimates of relative humidity (percentage)}

\item{tair}{vector of air temperatures (deg C)}

\item{tleaf}{vector of leaf temperatures (deg C)}

\item{tground}{ground surface temperature (deg C)}

\item{soilrh}{soil effective relative humidity (percentage)}
}
\value{
a vector of air temperatures (deg C)
}
\description{
Calculate relative humidity below canopy
}
