% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensibleheat.R
\name{rhohair}
\alias{rhohair}
\title{Calculate molar density of air}
\usage{
rhohair(tc, pk)
}
\arguments{
\item{tc}{air temperature (degrees C)}

\item{pk}{atmospheric pressure (kPa)}
}
\value{
molar density of air (kPa).
}
\description{
Calculate molar density of air
}
