% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow.R
\name{rossindex}
\alias{rossindex}
\title{Calculates slope correction for direct radiation using Ross G}
\usage{
rossindex(solp, y = 0.7, slope = 0, aspect = 0)
}
\arguments{
\item{solp}{a list of solar zenith and azimuths as returned by [sunposition()]}

\item{y}{parameter controlling angular response, where \code{y = 0} gives a
planar surface and \code{y = 1} gives isotropic response}

\item{slope}{surface slope (degrees)}

\item{aspect}{surface aspect (degrees)}
}
\value{
fraction of direct beam radiation incident on the inclined surface
}
\description{
Returns the fraction of direct beam radiation incident on an inclined
surface, analogous to [solarindex()], but with reduced angular dependence
controlled by \code{y}
}
