% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensibleheat.R
\name{roughlength}
\alias{roughlength}
\title{Calculate roughness length for momentum}
\usage{
roughlength(h, pai, d, psih)
}
\arguments{
\item{h}{vegetation height (m)}

\item{pai}{plant area index}

\item{psi_h}{diabatic correction for heat}
}
\value{
roughness length (m)
}
\description{
Calculate roughness length for momentum
}
