% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentheat.R
\name{satvap}
\alias{satvap}
\title{Calculate saturated vapour pressure}
\usage{
satvap(tc)
}
\arguments{
\item{tc}{temperature (degrees C)}
}
\value{
saturated vapour pressure (kPa).
}
\description{
Calculate saturated vapour pressure
}
