% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{scattercoef}
\alias{scattercoef}
\title{Scattering directionality factor}
\usage{
scattercoef(lref, ltra, x, method = "Sellers")
}
\arguments{
\item{lref}{leaf reflectance (0-1)}

\item{ltra}{leaf transmittance. Note that `ltra + lref` must be < 1.}

\item{method}{one of "Sellers" - leaf directionality factor approximated by
calculating mean leaf angle or "Pinty" - leaf directionality factor calculated
by integrating the inclination density distribution function.}
}
\value{
J - the scattering directionality factor
}
\description{
Calculates the canopy scattering directionality factor for use in two-stream models
using one of two methods
}
