% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow.R
\name{SnowAlbedo}
\alias{SnowAlbedo}
\title{Snow albedo as a function of age}
\usage{
SnowAlbedo(snowage)
}
\arguments{
\item{snowage}{Snow age (hours).}
}
\value{
Numeric value (or vector) of snow albedo (dimensionless, 0 to 1).
}
\description{
Computes broadband snow albedo as a function of snow age using
an empirical logarithmic decay relationship.
}
