% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow.R
\name{SnowDensFun}
\alias{SnowDensFun}
\title{Bulk snow density from depth and age}
\usage{
SnowDensFun(densfun, snowdepth, snowage)
}
\arguments{
\item{densfun}{Numeric vector of length 4 as returned by [SnowDensParams()].}

\item{snowdepth}{Snow depth (m).}

\item{snowage}{Snow age (hours).}
}
\value{
Numeric value of bulk snow density (kg m^-3).
}
\description{
Computes bulk snow density as a function of snow depth and
snow age using an empirical exponential formulation.
}
