% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow.R
\name{SnowDensParams}
\alias{SnowDensParams}
\title{Snow density parameter sets}
\usage{
SnowDensParams(snowenv = "Maritime")
}
\arguments{
\item{snowenv}{Character string specifying the snow environment. Options are:
`Alpine`, `Maritime`, `Prairie`, `Tundra` or `Taiga`.}
}
\value{
A numeric vector of length 4 containing density parameters:
\describe{
  \item{densfun[1]}{Upper asymptotic snow density (g cm^-3)}
  \item{densfun[2]}{Initial / minimum snow density (g cm^-3)}
  \item{densfun[3]}{Depth-dependent densification coefficient (m^-1)}
  \item{densfun[4]}{Age-dependent densification coefficient (day^-1)}
}
}
\description{
Returns empirical parameter values for a snow density evolution
function, based on different climatic snow regimes.

These parameterisations are intended for use with [snowdenfun()] to estimate
bulk snow density as a function of snow depth and age.
}
