% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow.R
\name{SnowHeatCapacity}
\alias{SnowHeatCapacity}
\title{Snow volumetric heat capacity}
\usage{
SnowHeatCapacity(rho, Tc, pk = 101.3)
}
\arguments{
\item{rho}{Snow density (kg m-3).}

\item{Tc}{Snow temperature (deg C).}

\item{pk}{Air pressure (kPa).}
}
\value{
Snow volumetric heat capacity (J m-3 K-1).
}
\description{
Calculates volumetric heat capacity of snow as a mixture of ice,
air, and optionally liquid water.
}
