% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow.R
\name{SnowResistance}
\alias{SnowResistance}
\title{Aerodynamic resistance above snow}
\usage{
SnowResistance(snowdepth, hgt, pai, uref, psim, psih, zref)
}
\arguments{
\item{snowdepth}{Snow depth (m).}

\item{hgt}{Vegetation height (m).}

\item{pai}{Plant area index (m^2 m^-2).}

\item{uref}{Wind speed at the reference height (m s^-1).}

\item{psim}{Stability correction term for momentum (dimensionless).}

\item{psih}{Stability correction term for heat (dimensionless).}

\item{zref}{Reference height above the ground surface (m).}
}
\value{
Aerodynamic resistance to heat transfer (\eqn{s m^{-1}}).
}
\description{
Computes aerodynamic resistance to heat transfer between a snow
surface and the reference height, accounting for partial burial of vegetation
by snow where relevant.
}
\details{
If snow partially buries the vegetation canopy, vegetation height and plant
area index are reduced in proportion to the exposed canopy above the snow
surface. Zero-plane displacement height and roughness length for momentum are
then recalculated for the reduced canopy using [zeroplanedis()] and
[roughlength()].
}
