% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow.R
\name{SnowSoilHeatModel}
\alias{SnowSoilHeatModel}
\title{Updates snow and soil temperatures using the coupled snow-soil heat model}
\usage{
SnowSoilHeatModel(
  snowmod,
  soilheatmod,
  soilp,
  Rabs,
  Tref,
  relhum,
  atmPressure,
  rHa,
  dT = 3600,
  Fact = 0.5,
  maxNrIterations = 100,
  tolerance = 0.01,
  relax = 0.3
)
}
\arguments{
\item{snowmod}{list of variables provided as input and updated by the snow
model as returned by [InitSnowModel()] on first step or by previous model run}

\item{soilheatmod}{list of variables provided as input and updated by soil
heat model as returned by [InitSoilheatmod()] on first step or by previous model run}

\item{soilp}{list of soil parameters as returned by [createsoilplist()]}

\item{Rabs}{absorbed radiation at the surface (W m-2)}

\item{Tref}{reference air temperature (degrees C)}

\item{relhum}{air relative humidity (Percentage)}

\item{atmPressure}{air pressure (kPa)}

\item{rHa}{aerodynamic resistance to heat from the surface to reference height
as returned by [SnowResistance()] or [groundresistance()] (s m-1)}

\item{dT}{time step (s)}

\item{Fact}{weighting factor for the implicit solution scheme}

\item{maxNrIterations}{maximum number of nonlinear iterations}

\item{tolerance}{convergence criterion for maximum change in layer temperature
between successive iterations (degrees C)}

\item{relax}{under-relaxation factor applied to temperature updates}
}
\value{
a list comprising
\describe{
  \item{snowmod}{updated snow model state}
  \item{soilheatmod}{updated soil heat model state}
}
}
\description{
Solves heat conduction through the combined snow and soil profile over one
time step, updating layer temperatures by combining surface energy balance from
[snowsurfaceEB()] or [soilsurfaceEB()], thermal conductivity from
[SnowThermalConductivity()] and [thermalConductivity()], volumetric heat capacity from
[SnowHeatCapacity()] and [heatCapacity()], and tridiagonal solution with
[ThomasBoundaryCondition()].
}
