% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow.R
\name{snowsurfaceEB}
\alias{snowsurfaceEB}
\title{Calculates snow surface energy balance}
\usage{
snowsurfaceEB(Rabs, Tref, Tsurface, pk, relhum, rHa)
}
\arguments{
\item{Rabs}{absorbed radiation at the surface (W m-2)}

\item{Tref}{reference air temperature (degrees C)}

\item{Tsurface}{snow surface temperature (degrees C)}

\item{pk}{air pressure (kPa)}

\item{relhum}{air relative humidity (Percentage)}

\item{rHa}{aerodynamic resistance to heat from snow surface to reference height
(s m-1), e.g. as returned by [SnowResistance()]}
}
\value{
snow surface energy balance (W m-2)
}
\description{
Calculates the snow surface energy balance (W m-2)
}
