% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snow.R
\name{SnowThermalConductivity}
\alias{SnowThermalConductivity}
\title{Snow thermal conductivity}
\usage{
SnowThermalConductivity(rho)
}
\arguments{
\item{rho}{Snow density (kg m-3).}
}
\value{
Snow thermal conductivity (W m-1 K-1).
}
\description{
Calculates snow thermal conductivity from bulk snow density
using an exponential relationship following Djachkova (after Anderson, 2006).
}
