% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{SoilHeatModel}
\alias{SoilHeatModel}
\title{Updates soil temperatures using the soil heat model}
\usage{
SoilHeatModel(
  soilheatmod,
  soilp,
  Rabs,
  Tref,
  relhum,
  atmPressure,
  rHa,
  dT = 3600,
  Fact = 0.5,
  maxNrIterations = 100,
  tolerance = 0.01
)
}
\arguments{
\item{soilheatmod}{list of variables provided as input and updated by soil
heat model as returned by [InitSoilheatmod()] on first step or by previous model run}

\item{soilp}{list of soil parameters as returned by [createsoilplist()]}

\item{Rabs}{absorbed radiation at the ground surface (W m-2)}

\item{Tref}{reference air temperature (degrees C)}

\item{relhum}{air relative humidity (Percentage)}

\item{atmPressure}{air pressure (kPa)}

\item{rHa}{aerodynamic resistance to heat from ground to reference height
as returned by [groundresistance()] (s m-1)}

\item{dT}{time step (s)}

\item{Fact}{weighting factor for the implicit solution scheme}

\item{maxNrIterations}{maximum number of nonlinear iterations}

\item{tolerance}{convergence criterion for maximum change in layer temperature
between successive iterations (degrees C)}
}
\value{
a list comprising
\describe{
  \item{n}{number of soil layers}
  \item{z}{depth of soil nodes (m)}
  \item{dz}{soil layer thicknesses (m)}
  \item{zCenter}{depth of layer centres (m)}
  \item{vol}{soil layer volumes per unit ground area (m)}
  \item{wc}{volumetric soil water content of each layer (m3 m-3)}
  \item{Te}{updated soil layer temperature (degrees C)}
  \item{oldTe}{soil layer temperature at previous time step (degrees C)}
  \item{Gflux}{integrated soil heat storage flux over the profile (W m-2)}
  \item{iters}{number of iterations used to reach convergence}
}
}
\description{
Solves soil heat conduction through the soil profile over one time step,
updating layer temperatures by combining surface energy balance from [soilsurfaceEB()],
thermal conductivity from [thermalConductivity()], volumetric heat capacity from
[heatCapacity()], and tridiagonal solution with [ThomasBoundaryCondition()].
}
