% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{soilrelhum}
\alias{soilrelhum}
\title{Calculates soil surface relative humidity}
\usage{
soilrelhum(theta, tsoil, psie, b, Smax)
}
\arguments{
\item{theta}{volumetric soil water content (m^3 / m^3)}

\item{tsoil}{soil temperature (degrees C)}

\item{psie}{air-entry water potential}

\item{b}{Campbell soil water retention parameter}

\item{Smax}{saturated volumetric water content (m^3 / m^3)}
}
\value{
relative humidity (Percentage)
}
\description{
Calculates soil surface relative humidity
}
