% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{soilsurfaceEB}
\alias{soilsurfaceEB}
\title{Calculates soil surface energy balance}
\usage{
soilsurfaceEB(soilp, Rabs, Tref, Tsurface, pk, relhum, rHa, theta)
}
\arguments{
\item{soilp}{list of soil parameters as returned by [createsoilplist()]}

\item{Rabs}{absorbed radiation at the surface (W m-2)}

\item{Tref}{reference air temperature (degrees C)}

\item{Tsurface}{soil surface temperature (degrees C)}

\item{pk}{air pressure (kPa)}

\item{relhum}{air relative humidity (Percentage)}

\item{rHa}{aerodynamic resistance to heat from ground to reference height
as returned by [groundresistance()] (s m-1)}

\item{theta}{volumetric soil water content (m3 m-3)}
}
\value{
soil surface energy balance (W m-2)
}
\description{
Calculates the soil surface energy balance (W m-2)
}
