% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{solarindex}
\alias{solarindex}
\title{Calculates solar index}
\usage{
solarindex(slope, aspect, solp)
}
\arguments{
\item{slope}{slope (decimal degrees from horizontal)}

\item{aspect}{aspect (decimal degrees from north)}

\item{solp}{a list of solar zenith and azimuths as returned by [sunposition()]}
}
\value{
fraction of direct beam radiation incident on
an inclined surface
}
\description{
Calculates fraction of direct beam radiation incident on
an inclined surface
}
