% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance.R
\name{solve_wholecanopy}
\alias{solve_wholecanopy}
\title{Solve energy balance equation for entire canopy returning associated variables}
\usage{
solve_wholecanopy(
  weather,
  vegp,
  groundp,
  hr,
  lat,
  long,
  zref,
  G,
  theta,
  soilrh,
  Ca = 430,
  maxiter = 20
)
}
\arguments{
\item{weather}{a data.frame of weather variables matching the format of `micropoint::climdata`}

\item{vegp}{a list of vegetation parameters as returned by [createplant_inputs()]}

\item{groundp}{a list of ground parameters matching the format of `micropoint::groundparams`}

\item{hr}{hour in weather for energy balance is to be solved}

\item{lat}{latitude (decimal degrees)}

\item{long}{longitude (decimal degrees)}

\item{zref}{height above ground of weather variables (must be above canopy - see details)}

\item{G}{Rate of heat storage by ground at `hr` (W/m^2)}

\item{theta}{soil water fraction in root zone}

\item{soilrh}{effective soil relative humidity (percentage)}

\item{Ca}{Atmospheric CO2 concentration in ppm as returned by `micropoint::Cafromyear()`}

\item{maxiter}{Maxiumum number of iterations to achieve convergence}
}
\value{
a list of of the following variables
\describe{
 \item{tcanopy}{temperature of canopy heat exchange surface (deg C)}
 \item{Rabs}{Flux density of radiation absorbed by canopy (W/m^2)}
 \item{H}{Flux density of sensible heat exchanged between canopy and `zref` (W/m^2)}
 \item{L}{Flux density of latent heat exchanged between canopy and `zref` (W/m^2)}
 \item{uf}{wind friction velocity (m/s)}
 \item{uh}{wind speed at top of canopy (m/s)}
 \item{th}{temperature at top of canopy (deg C)}
 \item{rHa}{bulk surface aerodynamic resistance (s/m)}
 \item{rSl}{bulk surface stomatal resistance (s/m)}
 \item{rS}{bulk surface resistance to vapour including from ground surface (s/m)}
 \item{psim}{diabatic influencing factor for momentum}
 \item{psih}{diabatic influencing factor for heat}
 \item{LL}{Obukhov length}
 \item{TL}{Langrangian time-scale}
}
}
\description{
Solve energy balance equation for entire canopy returning associated variables
}
\details{
Meteorological inputs must represent above-canopy conditions,
whereas forcing data are often measured 1–2 m above the ground, below the
height of surrounding vegetation (e.g. forest canopy). The function
`weatherhgt_adjust` in the `micropoint` package  can be used to adjusts
weather variables to estimate the values that would occur sufficiently high
above the surface, assuming site characteristics consistent with WMO guidelines
for weather station siting.
}
