% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energybalance.R
\name{SolveEnergyBalance}
\alias{SolveEnergyBalance}
\title{Solve energy balance equation to derive surface temperature}
\usage{
SolveEnergyBalance(
  Ta,
  pk,
  rh,
  Rabs,
  rHa,
  rS,
  G,
  em = 0.97,
  method = c("uniroot", "Penman"),
  iters = 1
)
}
\arguments{
\item{Ta}{Air temperature (degrees Celsius).}

\item{pk}{Atmospheric pressure (kPa).}

\item{rh}{Relative humidity (percentage).}

\item{Rabs}{Flux density of absorbed radiation (W/m^2).}

\item{rHa}{Resistance to heat loss (s/m).}

\item{rS}{Stomatal resistance (s/m).}

\item{G}{Rate of heat storage (W/m^2).}

\item{em}{Emissivity of surface (unitless). Default is 0.97.}

\item{method}{One of `'uniroot'` or `'Penman'` (see details for more information on each method).}

\item{iters}{Integer specifying the number of iterations for the Penman method. Default is 1.}
}
\value{
A numeric vector of surface temperatures (in degrees Celsius) corresponding to each input
         air temperature (`Ta`).
}
\description{
This function solves for the surface (or organism) temperature (\eqn{T_s})
that satisfies the steady-state energy balance, based on various meteorological
and physical parameters. The energy balance can be solved using either the
**uniroot method** (slower) or the **Penman-Monteith method** (faster).
Users also have the option to iterate the Penman method for a more accurate solution.
}
\details{
This function solves for the surface (or organism) temperature (\eqn{T_s})
that satisfies the steady-state energy balance, based on various meteorological
and physical parameters. The energy balance can be solved using either the
**uniroot method** (slower) or the **Penman-Monteith method** (faster).
Users also have the option to iterate the Penman method for a more accurate solution.
}
