% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentheat.R
\name{stomatalcond_calc}
\alias{stomatalcond_calc}
\title{Calculate leaf stomatal conductance}
\usage{
stomatalcond_calc(
  Ca,
  Rswabs,
  tair,
  tleaf,
  rh,
  pk,
  psi_r,
  plant_inputs,
  z,
  C3 = TRUE
)
}
\arguments{
\item{Ca}{atmospheric carbon dioxide concentration (ppm)}

\item{Rswabs}{flux density of PAR absorbed by leaf (W/m^2)}

\item{tair}{temperature of air (deg C)}

\item{tleaf}{temperature of leaf (deg C)}

\item{rh}{relative humidity of air (percentage)}

\item{pk}{atmospheric pressure (kPa)}

\item{psi_r}{# mean water potential in root zone (MPa)}

\item{plant_inputs}{a vector of vegetation parameters as returned by [createplant_inputs()]}

\item{z}{height above ground}

\item{C3}{optional logical indicating whether vegetation has C3 or C4 photosynthetic pathway}
}
\value{
stomatal conductance (mol / m^2 / s)
}
\description{
Calculate leaf stomatal conductance
}
