% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{sunposition}
\alias{sunposition}
\title{Calculates solar position}
\usage{
sunposition(tme, lat, long)
}
\arguments{
\item{tme}{POSIXlt object of date and time in UTC}

\item{lat}{latitude (decimal degrees)}

\item{long}{longitude(decimal degrees, -ve west of Greenwich meridian)}
}
\value{
a list comprising
\describe{
 \item{zen}{the solar zenith angle (degrees from vertical)}
 \item{azi}{the solar azimuth angle (degrees from north)}
}
}
\description{
Calculates the solar zenith and azimuth angles
}
