% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensibleheat.R
\name{temp_below}
\alias{temp_below}
\title{Calculate air temperature below canopy}
\usage{
temp_below(hgt, paii, TL, uf, pk, Hz, tair, tleaf, tground)
}
\arguments{
\item{hgt}{height of canopy(m)}

\item{paii}{vector of plant area index values per unit ground area ordered
sequentially from canopy bottom to top.}

\item{TL}{Langrangian timescale as returned by [solve_wholecanopy()]}

\item{uf}{wind friction velocity as returned by [solve_wholecanopy()]}

\item{pk}{atmospheric pressure (kPa)}

\item{Hz}{vector of sensible heat fluxes from each canopy layer (W/m^2)}

\item{tair}{vector of initial estimates of air temperatures (deg C).}

\item{tleaf}{vector of leaf temperatures (deg C)}

\item{tground}{ground surface temperature (deg C)}
}
\value{
a vector of air temperatures (deg C)
}
\description{
Calculate air temperature below canopy
}
