% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{thermalConductivity}
\alias{thermalConductivity}
\title{Calculates soil thermal conductivity}
\usage{
thermalConductivity(Vq, Vm, Vo, Vw, Mc, Tc, pk)
}
\arguments{
\item{Vq}{volumetric quartz fraction of soil (m^3 / m^3)}

\item{Vm}{volumetric mineral fraction of soil (m^3 / m^3)}

\item{Vo}{volumetric organic fraction of soil (m^3 / m^3)}

\item{Vw}{volumetric water content of soil (m^3 / m^3)}

\item{Mc}{mass fraction of clay (kg / kg)}

\item{Tc}{soil temperature (degrees C)}

\item{pk}{air pressure (kPa)}
}
\value{
soil thermal conductivity (W m-1 K-1)
}
\description{
Calculates soil thermal conductivity (W m-1 K-1)
}
