% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{thetaFromPsi}
\alias{thetaFromPsi}
\title{Calculates volumetric soil water fraction from water potential using the Campbell method}
\usage{
thetaFromPsi(psiw, psie, b, Smax)
}
\arguments{
\item{psiw}{soil water potential (J kg-1)}

\item{psie}{air-entry water potential (J kg-1)}

\item{b}{Campbell soil water retention parameter}

\item{Smax}{volumetric soil water content at saturation (m3 m-3)}
}
\value{
volumetric soil water content (m3 m-3)
}
\description{
Calculates volumetric soil water fraction from water potential using the Campbell method
}
