% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{thetaFromPsiVG}
\alias{thetaFromPsiVG}
\title{Calculates volumetric soil water fraction from water potential using the Van Genuchten method}
\usage{
thetaFromPsiVG(psiw, Smin, Smax, psie, alpha, n)
}
\arguments{
\item{psiw}{soil water potential (kPa ~ J/Kg)}

\item{Smin}{Residual volumetric soil water fraction (m^3 / m^3)}

\item{Smax}{volumetric soil water fraction at saturation (m^3 / m^3)}

\item{psie}{air entry water potential (J/m^3)}

\item{alpha}{Van Genuchten shape parameter}

\item{n}{Van Genuchten pore size distribution parameter}
}
\value{
volumetric soil water content (m3 m-3)
}
\description{
Calculates volumetric soil water fraction from water potential using the Van Genuchten method
}
