% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{ThomasBoundaryCondition}
\alias{ThomasBoundaryCondition}
\title{Solves a tridiagonal system of equations}
\usage{
ThomasBoundaryCondition(aa, bb, cc, dd, x, first = 1, last = length(x))
}
\arguments{
\item{aa}{vector of sub-diagonal coefficients}

\item{bb}{vector of diagonal coefficients}

\item{cc}{vector of super-diagonal coefficients}

\item{dd}{vector of right-hand-side values}

\item{x}{vector of solution values}

\item{first}{index of first equation to solve}

\item{last}{index of last equation to solve}
}
\value{
a list comprising
\describe{
  \item{bb}{updated diagonal coefficients}
  \item{cc}{updated super-diagonal coefficients}
  \item{dd}{updated right-hand-side values}
  \item{x}{solution vector}
}
}
\description{
Solves a tridiagonal system of linear equations using the Thomas algorithm
}
