% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{twostream}
\alias{twostream}
\title{Two-stream radiation model}
\usage{
twostream(vegp, groundp, paii, swdown, difrad, lat, long, solp, PAR = FALSE)
}
\arguments{
\item{vegp}{a list of vegetation parameters - see e.g. example dataset micropoint::vegparams}

\item{groundp}{a list of ground parameters - see e.g. micropoint::groundparams}

\item{paii}{vector of plant area index values per unit ground area ordered
sequentially from canopy bottom to top. Note sum(paii) should equal vegp$pai}

\item{swdown}{a single numeric value of flux density of downward shortwave
radiation on a flat horizontal surface}

\item{difrad}{a single numeric value of flux density of downward diffuse radiation}

\item{lat}{latitude (decimal degrees), negative south of the equator}

\item{long}{longitude (decimal degrees), negative west of the Greenwich meridian}

\item{solp}{a list of the solar zenith and azimith angle}
}
\value{
a list of radiation fluxes in the same units as `swrad` and `difrad`
\describe{
 \item{Rdirdown}{Flux density of downward direct radiation perpendicular to the solar beam}
 \item{Rdifdown}{Flux density of downward diffuse radiation}
 \item{Rswup}{Flux density of upward shortwave radiation - assumed entirely diffuse}
 \item{Rleafabs}{Flux density of shortwave radiation absorbed by leaves - averaged over both sides}
 \item{Rsun}{Flux density of shortwave radiation absorbed by sunlit leaves}
 \item{Rshade}{Flux density of shortwave radiation absorbed by shaded leaves}
 \item{sunlitfrac}{Fraction of leaves that are sunlit}
 \item{Rgroundabs}{Flux density of shortwave radiation absorbed by ground surface}
}
}
\description{
Calculates upward and downward fluxes of shortwave radiation within canopies
accomodating inclined surfaces and non-random canopy gaps
}
