% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{vaporConductivityFromPsiTheta}
\alias{vaporConductivityFromPsiTheta}
\title{Calculates vapour conductivity from soil water potential}
\usage{
vaporConductivityFromPsiTheta(psiw, theta, Tk, Smax)
}
\arguments{
\item{psiw}{soil water potential (J kg-1)}

\item{theta}{volumetric soil water content (m3 m-3)}

\item{Tk}{soil temperature (K)}

\item{Smax}{volumetric soil water content at saturation (m3 m-3)}
}
\value{
soil vapour conductivity (kg^2 m-1 s-1 J-1)
}
\description{
Calculates soil vapour conductivity as a function of water potential and moisture content
}
