% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{vaporFromPsi}
\alias{vaporFromPsi}
\title{Calculates soil vapour content from water potential}
\usage{
vaporFromPsi(psiw, theta, Tk, Smax)
}
\arguments{
\item{psiw}{soil water potential (J kg-1)}

\item{theta}{volumetric soil water content (m3 m-3)}

\item{Tk}{soil temperature (K)}

\item{Smax}{volumetric soil water content at saturation (m3 m-3)}
}
\value{
soil vapour content
}
\description{
Calculates soil vapour content from soil water potential using the Kelvin equation
}
