% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground.R
\name{VGparams}
\alias{VGparams}
\title{Create list of Van Genuchten soil hydraulic parameters}
\usage{
VGparams(soiltype)
}
\arguments{
\item{soiltype}{Character string giving the soil type. See details}
}
\value{
A list of soil parameters:
\describe{
  \item{psie}{Air-entry water potential (m)}
  \item{alpha}{Van Genuchten shapte parameter (unitless)}
  \item{n}{Pore size distribution parameter (unitless)}
  \item{Smax}{Volumetric soil water content at saturation (m^3 m^-3)}
  \item{Smin}{Residual volumetric soil water content (m^3 m^-3)}
  \item{Ksat}{Saturated hydraulic conductivity (kg s / m^3)}
}
}
\description{
Create list of Van Genuchten soil hydraulic parameters
}
\details{
:
Available soil types (argument \code{soiltype}) are:
\itemize{
  \item Sand
  \item Loamy sand
  \item Sandy loam
  \item Loam
  \item Silt loam
  \item Sandy clay loam
  \item Clay loam
  \item Silty clay loam
  \item Sandy clay
  \item Silty clay
  \item Clay
  \item Silt
}
}
