% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensibleheat.R
\name{windfric}
\alias{windfric}
\title{Calculate wind friction velocity}
\usage{
windfric(uref, zref, d, zm, psim)
}
\arguments{
\item{uref}{wind speed at reference height (m/s)}

\item{zref}{height above ground of wind speed measurement (m)}

\item{d}{zero-plane displacement height as returned by [zeroplanedis()] (m)}

\item{zm}{roughness length for momentum as returned by [roughlength()] (m)}

\item{psim}{diabatic correction for momentum as returned by [dpsim()]}
}
\value{
friction velocity (m/s)
}
\description{
Calculate wind friction velocity
}
