% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensibleheat.R
\name{windprofile_above}
\alias{windprofile_above}
\title{Calculate wind profile above canopy}
\usage{
windprofile_above(z, uf, d, zm, psim)
}
\arguments{
\item{z}{height above ground for which wind speed desired (m)}

\item{uf}{wind friction velocity as returned by [windfric()] (m/s)}

\item{d}{zero-plane displacement height as returned by [zeroplanedis()] (m)}

\item{zm}{roughness length for momentum as returned by [roughlength()] (m)}

\item{psim}{diabatic correction for momentum as returned by [dpsim()]}
}
\value{
wind speed (m/s)
}
\description{
Calculate wind profile above canopy
}
