% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensibleheat.R
\name{windprofile_below}
\alias{windprofile_below}
\title{Calculate wind profile above canopy}
\usage{
windprofile_below(hgt, paii, uh)
}
\arguments{
\item{hgt}{height of canopy(m)}

\item{paii}{vector of plant area index values per unit ground area ordered
sequentially from canopy bottom to top.}

\item{uh}{wind speed at top of canopy as derived using [windprofile_above()] (m/s)}
}
\value{
wind speed (m/s)
}
\description{
Calculate wind profile above canopy
}
