% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensibleheat.R
\name{zeroplanedis}
\alias{zeroplanedis}
\title{Calculate zero-plane displacement height}
\usage{
zeroplanedis(h, pai)
}
\arguments{
\item{h}{vegetation height (m)}

\item{pai}{plant area index}
}
\value{
zero-plane displacement height (m)
}
\description{
Calculate zero-plane displacement height
}
